+++ packages/fhi-designsystem/.storybook/preview.ts
@@ -12,7 +12,7 @@ const preview: Preview = {
     },
     options: {
       storySort: {
-        order: ['*', 'Ikoner', 'Design Tokens', ['Introduksjon']],
+        order: ['Kom i gang', '*', 'Ikoner', 'Design Tokens', ['Introduksjon']],
       },
     },
     designToken: {